Move GdgNavDrawerActivity to common package
