+++ app/src/main/java/org/gdg/frisbee/android/common/GdgNavDrawerActivity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gdg.frisbee.android.activity;
+package org.gdg.frisbee.android.common;
 
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -44,6 +44,8 @@ import org.gdg.frisbee.android.BuildConfig;
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
 import org.gdg.frisbee.android.about.AboutActivity;
+import org.gdg.frisbee.android.activity.GdgActivity;
+import org.gdg.frisbee.android.activity.SettingsActivity;
 import org.gdg.frisbee.android.adapter.DrawerAdapter;
 import org.gdg.frisbee.android.adapter.DrawerAdapter.DrawerItem;
 import org.gdg.frisbee.android.api.GapiOkTransport;
