+++ app/src/main/java/org/gdg/frisbee/android/eventseries/TaggedEventSeriesActivity.java
@@ -25,7 +25,7 @@ import android.widget.TextView;
 
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
-import org.gdg.frisbee.android.activity.GdgNavDrawerActivity;
+import org.gdg.frisbee.android.common.GdgNavDrawerActivity;
 
 import butterknife.InjectView;
 import butterknife.Optional;
